requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cfgv==3.3.1
charset-normalizer==3.1.0
colorama==0.4.6
colorlog==6.7.0
distlib==0.3.6
filelock==3.12.0
frozenlist==1.3.3
hikari==2.0.0.dev119
hikari-lightbulb==2.3.3
identify==2.5.24
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
nodeenv==1.8.0
packaging==23.1
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0
virtualenv==20.23.0
yarl==1.9.2
